requirements.txt
psycopg2-binary>=2.9.3
xxhash==3.3.0
python-dotenv>=0.20.0
tabulate>=0.8.10
colorama>=0.4.4
